refactor(apps/web): extract queue components into separate modules
Extract queue-related UI logic from the main Queue component into
focused, reusable subcomponents: QueueHeader, QueueNowPlaying,
QueueNextUp, and QueueItem. This improves maintainability and enables
independent animation handling for each section.

- Move header controls to QueueHeader component
- Extract now playing section to QueueNowPlaying with motion animations
- Extract next up list to QueueNextUp with drag-and-drop and animations
- Extract individual queue item to QueueItem with motion layout support
- Add framer-motion animations for smooth transitions and list reordering
- Rename QueueItem type import to PlayrQueueItem to avoid naming conflicts
